bring up to date; address comments.
